Add Collection view can load meshes and textures from a remote server
CollectionObjCollectionLoad.js now loads files remotely using ObjReader's ConvertFileAsync method, which
handles downloading in addition to converting an obj to a GameObject.

Preprocessor directives have been used to ensure that the Unity Editor and WebGL builds target remote
files, and that the standalone build targets local files. It will be necessary to change these directives
when testing standalone vs. WebGL.

Note that this commit does not completely implement remote file loading for WebGL -- the WebGL build is
undermined by the XML library's propensity to spawn threads, which is not supported by web browsers.