Refactor XML download out of CollectionReader
The CollectionReader class was previously responsible for downloading XML data, violating the
single responsibility principle (i.e. a CollectionReader need only know how to interpret XML data
as Collection information; it shouldn't be responsible for managing IO)

The new API requires the user of the CollectionReader to download some XML, then call CollectionReader.LoadXmlFromText(string xmlText)
and pass in a textual representation of the XML data. The remainder of the API is unchanged.

See Collection_AvailableCollections.DownloadXML for a suggested use of the CollectionReader