Refactor XML download out of DublinCoreReader
The DublinCoreReader now requires the caller to populate it with XML data. In general, this should be
done by checking whether the Reader requires data (via DublinCoreReader.HasData()); if not, download
some XML and pass it in to the Reader as text (DublinCoreReader.LoadXmlFromText(string text)).

Note that the redistribution of the responsibility to download data has only been incorporated for the
Collection scene (c.f Collect_CollectControl.cs). Other classes requiring a change have been marked with
TODO.